// src/types.ts
export type User = firebase.User;

export type UserRole = 'admin' | 'general_manager' | 'area_manager' | 'store_manager' | 'employee';

export interface UserProfile {
    id: string; // Will be the user's UID
    name: string;
    role: UserRole;
    email: string;
    employeeId?: string;
    phone?: string;
    areaManager?: string; // For area_manager role
    store?: string; // For store_manager and employee roles
    status?: 'pending' | 'active';
}